Refs #19803: address windows warning

+++ test/blackbox/common/BlackboxTestsPubSubHistory.cpp
@@ -992,7 +992,7 @@ TEST_P(PubSubHistory, PubSubAsReliableKeepLastWithKeyUnorderedReception)
     writer.send(data);
     ASSERT_TRUE(data.empty());
 
-    reader.block_for_at_least(keys * depth * 0.1);
+    reader.block_for_at_least(static_cast<size_t>(keys * depth * 0.1));
 
     //! Avoid dropping determiniscally the same re-sent samples
     testTransport->dropDataMessagesPercentage.store(10);
